requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0
django-filter==1.1.0
django-heroku==0.3.1
django-rest-swagger==2.1.2
djangorestframework==3.8.2
drf-extensions==0.3.1
drf-nested-routers==0.90.2
drf-yasg==1.7.1
flex==6.13.1
future==0.16.0
gunicorn==19.8.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
permission==0.4.1
psycopg2==2.7.4
Pygments==2.2.0
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rfc3987==1.3.7
ruamel.yaml==0.15.37
simplejson==3.13.2
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
whitenoise==3.3.1
